Provide more information in the Auth0 Slack alerts
We don't send the description from the log event to Slack because it
might contain PII, but there is a "log event type code" that we can map
to a human-readable description.

This should make triage of events slightly easier, by allowing us to
make an initial assessment before we open the dashboard.