Separate the Lambda that alerts on us-east-1 Lambda errors
The issue with Terraform wanting to update IAM permissions indicates a
bug in the way we've set up these Lambdas; in particular:

*   We have a Lambda to catch Lambda errors in experience/eu-west-1
    (which is where most of our Lambdas live)

*   We have a Lambda to catch Lambda errors in experience/us-east-1
    (which is where our CloudFront Lambdas live)

This Lambda includes an IAM role which allows them to write to their own
DLQ (so they can report on their own errors), and IAM roles are globally
unique.  Before this patch, both Lambdas had roles with the same name,
so they were fighting over whose queue they'd be allowed to write to.

This tweaks the names so the Lambda in us-east-1 has a different name,
which means it has a different IAM role.  This means they're no longer
blatting each other's changes, and Terraform plans cleanly.