Upgrade the account.wellcomecollection.org CloudFront Terraform to the 4.x provider
There are no S3 resources so no deprecation warnings to fix, but there
is a local that points to a non-existent output in another stack.  Since
this is only used in commented-out code, I've commented out the local
for now -- it can be fixed by anybody who wants to use the commented-out
code.