Remove min-age on deletion phase of apm_traces ILM policies
This change allows indices rolled over hot to warm to be kept for 1 day, and then deleted straight away. Currently the hot phase rollover rules keep the apm_traces indices (these tend to be the largest) around for 30 days or until they reach 50GB (total disk size at present is 250GB).

The expected process is:

- After either 30 days, or the hot index reaches 30gb it is rolled over to the warm phase
- After another 1 day, the warm index (or indices if the hot index has rolled multiple times) rolls over to the delete phase
- The deletion can occur immediately, the default value for min_age is 0 days

If we do not have a roll-over from hot to warm we risk a situation where a hot index is immediately deleted with no buffer when it rolls over. Including the warm phase does increase the chance that the disk will fill up ... we're essentially betting we won't have multiple roll-overs within the space of day. We reduce the max size of the hot index from 50gb to 30gb to allow expected space for 1 hot index and 1 warm index, with the warm index providing us with a buffer before deletion.

For context see:
- https://www.elastic.co/guide/en/elasticsearch/reference/current/ilm-index-lifecycle.html#ilm-phase-transitions
- https://www.elastic.co/guide/en/elasticsearch/reference/current/ilm-rollover.html#_rollover_condition_blocks_phase_transition